# Backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chit Fund Guardian System API",
    "version": "1.0.0",
    "description": "API documentation for the Chit Fund Guardian System backend."
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
    "/user/create": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "email": { "type": "string" },
                  "phoneNo": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["firstName", "email", "phoneNo", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/user/login": {
      "post": {
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "400": { "description": "Validation error" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/user/request-password-reset": {
      "post": {
        "summary": "Request password reset (send OTP)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" }
                },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OTP sent if user exists" }
        }
      }
    },
    "/user/reset-password": {
      "post": {
        "summary": "Reset password with OTP",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "otp": { "type": "string" },
                  "newPassword": { "type": "string" }
                },
                "required": ["email", "otp", "newPassword"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Password reset successful" },
          "400": { "description": "Invalid or expired OTP" }
        }
      }
    },
    "/group/create": {
      "post": {
        "summary": "Create a new group (admin only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "chitValue": { "type": "number" },
                  "tenure": { "type": "integer" },
                  "startMonth": { "type": "string", "format": "date" },
                  "foremanCommission": { "type": "number" }
                },
                "required": ["chitValue", "tenure", "startMonth", "foremanCommission"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Group created" },
          "400": { "description": "Validation error" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/request/approve": {
      "post": {
        "summary": "Approve a user request (admin only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": { "type": "string" },
                  "seclectedgroupId": { "type": "string" },
                  "groupId": { "type": "string" }
                },
                "required": ["requestId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Request approved" },
          "400": { "description": "Validation error" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/user/audit-logs": {
      "get": {
        "summary": "View audit logs (admin only)",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Audit logs returned" },
          "403": { "description": "Forbidden" }
        }
      }
    }
  }
} 